moved socket logic into its own context module
