+++ components/Spaces/CallTabs/index.js
@@ -4,6 +4,7 @@ import { TabList, Tab, Tabs, TabPanel, resetIdCounter } from 'react-tabs';
 import { IconButton, Grid, Paper } from '@material-ui/core';
 import { Chat as ChatIcon, People as PeopleIcon, LibraryMusic, PlaylistAddCheck } from '@material-ui/icons';
 
+import { useSocketContext } from '@/context/spaces/SocketContext';
 import renderComponent from '@/utils/renderComponent';
 import TabPanelHeader from './Layout/TabPanelHeader';
 import Music from './Music';
@@ -14,7 +15,9 @@ import TodoList from './TodoList';
 // https://github.com/reactjs/react-tabs#api
 resetIdCounter();
 
-function CallTabs({ username, participants, socketRef, roomID, conversation, showTabs, setShowTabs }) {
+function CallTabs({ roomID, showTabs, setShowTabs }) {
+  const { username, participants, socketRef, conversation } = useSocketContext();
+
   useEffect(() => {
     setShowTabs(true);
   }, [setShowTabs]);
@@ -112,11 +115,7 @@ function CallTabs({ username, participants, socketRef, roomID, conversation, sho
 }
 
 CallTabs.propTypes = {
-  username: PropTypes.string.isRequired,
-  participants: PropTypes.array.isRequired,
-  socketRef: PropTypes.object.isRequired,
   roomID: PropTypes.string.isRequired,
-  conversation: PropTypes.array.isRequired,
   showTabs: PropTypes.bool.isRequired,
   setShowTabs: PropTypes.func.isRequired,
 };
