+++ components/Spaces/VideoOptions/CallOptions.js
@@ -1,5 +1,4 @@
 import { useState, useRef } from 'react';
-import PropTypes from 'prop-types';
 import { IconButton, MenuList, MenuItem, Popper, Grow, Paper, ClickAwayListener } from '@material-ui/core';
 import {
   Mic,
@@ -13,10 +12,12 @@ import {
   ExitToApp,
 } from '@material-ui/icons';
 
+import { useSocketContext } from '@/context/spaces/SocketContext';
 import LeaveCallDialog from './LeaveCallDialog';
 import ParticipantsDialog from './ParticipantsDialog';
 
-function CallOptions({ userAudioShow, toggleUserAudio, userVideoShow, toggleUserVideo, leaveCall }) {
+function CallOptions() {
+  const { enableUserAudio, enableUserVideo, toggleUserAudio, toggleUserVideo, leaveCall } = useSocketContext();
   const [openOptions, setOpenOptions] = useState(false);
   const [openLeaveModal, setOpenLeaveModal] = useState(false);
   const [openParticipantsModal, setOpenParticipantsModal] = useState(false);
@@ -37,8 +38,8 @@ function CallOptions({ userAudioShow, toggleUserAudio, userVideoShow, toggleUser
   return (
     <>
       <div className="absolute bottom-0 left-0">
-        <IconButton onClick={toggleUserAudio}>{userAudioShow ? <Mic /> : <MicOff />}</IconButton>
-        <IconButton onClick={toggleUserVideo}>{userVideoShow ? <Videocam /> : <VideocamOff />}</IconButton>
+        <IconButton onClick={toggleUserAudio}>{enableUserAudio ? <Mic /> : <MicOff />}</IconButton>
+        <IconButton onClick={toggleUserVideo}>{enableUserVideo ? <Videocam /> : <VideocamOff />}</IconButton>
         <IconButton onClick={handleToggle} ref={buttonRef}>
           <MoreVert />
         </IconButton>
@@ -78,13 +79,5 @@ function CallOptions({ userAudioShow, toggleUserAudio, userVideoShow, toggleUser
     </>
   );
 }
-
-CallOptions.propTypes = {
-  userAudioShow: PropTypes.bool.isRequired,
-  toggleUserAudio: PropTypes.func.isRequired,
-  userVideoShow: PropTypes.bool.isRequired,
-  toggleUserVideo: PropTypes.func.isRequired,
-  leaveCall: PropTypes.func.isRequired,
-};
 
 export default CallOptions;
